Add: 카페 좋아요 기능 추가

+++ likes/views.py
@@ -1,22 +1,26 @@
-from django.http import JsonResponse
+from django.http  import JsonResponse
 from django.views import View
 
 from cafes.models import Cafe
 from likes.models import CafeLike
+from utils import LoginConfirm
 
 class CafeLikeView(View):
     @LoginConfirm
     def post(self, request, cafe_id):
+        if not Cafe.objects.filter(id=cafe_id).exists:
+            return JsonResponse({'MESSAGE' : 'CAFE_DOES_NOT_EXIST'})
+
         if not CafeLike.objects.filter(cafe_id=cafe_id, user_id=request.user.id):
             CafeLike.objects.create(
                 user = request.user,
                 cafe = Cafe.objects.get(id=cafe_id)
             )
 
-            return JsonResponse({'MESSAGE' : 'CREATED'}, status=200)
+            return JsonResponse({'MESSAGE' : 'LIKE_CREATED'}, status=200)
 
         else:
             CafeLike.objects.filter(cafe_id=cafe_id, user_id=request.user.id).delete()
 
-            return JsonResponse({'MESSAGE' : 'DELETED'}, status=200)
+            return JsonResponse({'MESSAGE' : 'LIKE_DELETED'}, status=200)
